style: 配置文件代码优化
